chore: move types folder inside src directory + update imports
