+++ src/app/(client)/productos/page.tsx
@@ -3,7 +3,7 @@ import { Card, Filters, Loader, Pagination } from "@/components";
 import { useEffect, useState } from "react";
 import { QueryClient, QueryClientProvider, useQuery } from "react-query";
 import { getProducts } from "../../../sanity/requests/sanity-requests";
-import { Product } from "../../../../types/Product";
+import { Product } from "../../../types/Product";
 import { useSearchParams } from "next/navigation";
 import { useRouter } from "next/navigation";
 import { navigation } from "@/utils/navigation";
